Move horizon.go to the same package as channel_accounts_service.go
