+++ internal/transactionsubmission/services/horizon.go
@@ -1,4 +1,4 @@
-package transactionsubmission
+package services
 
 import (
 	"context"
