+++ internal/transactionsubmission/services/horizon_test.go
@@ -1,4 +1,4 @@
-package transactionsubmission
+package services
 
 import (
 	"context"
